consolidate into spec/ folder
