+++ src/CMakeLists.txt
@@ -87,6 +87,7 @@ file(GLOB SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/module-init/*.c"
     "${CMAKE_CURRENT_SOURCE_DIR}/obfuscation/*.c"
     "${CMAKE_CURRENT_SOURCE_DIR}/profile/*.c"
+    "${CMAKE_CURRENT_SOURCE_DIR}/spec/*.c"
 
     "${root}/deps/cndict/cndict_data.c"
     "${root}/deps/libnu/*.c"
