+++ src/info/info_command.h
@@ -8,7 +8,7 @@
 */#pragma once
 
 #include "redismodule.h"
-#include "spec.h"
+#include "spec/spec.h"
 
 #ifdef __cplusplus
 extern "C" {
