+++ src/notifications.c
@@ -8,7 +8,7 @@
 */
 #include "config.h"
 #include "notifications.h"
-#include "spec.h"
+#include "spec/spec.h"
 #include "doc_types.h"
 #include "redismodule.h"
 #include "rdb.h"
