+++ src/rdb.c
@@ -6,7 +6,7 @@
  * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
  * GNU Affero General Public License v3 (AGPLv3).
 */
-#include "spec.h"
+#include "spec/spec.h"
 #include "alias.h"
 #include "rdb.h"
