+++ src/redis_index.h
@@ -13,7 +13,7 @@
 #include "inverted_index.h"
 #include "search_ctx.h"
 #include "concurrent_ctx.h"
-#include "spec.h"
+#include "spec/spec.h"
 #include "iterators/iterator_api.h"
 
 #ifdef __cplusplus
