+++ src/spec/spec_info.c
