+++ src/spec/spec_info.h
