+++ src/spec/spec_rdb.c
