+++ src/spec/spec_rdb.h
