+++ src/spec/spec_registry.h
