+++ src/spec/spec_scanner.c
