+++ src/spec/spec_scanner.h
