+++ tests/cpptests/test_cpp_agg.cpp
@@ -13,7 +13,7 @@
 #include "redismock/redismock.h"
 #include "redismock/util.h"
 #include "redismock/internal.h"
-#include "spec.h"
+#include "spec/spec.h"
 #include "common.h"
 #include "module.h"
 #include "version.h"
